dev/package/requirements.txt
pyyaml  == 6.0.1
pytz    == 2024.1
boto3   == 1.34.49
slack_sdk == 3.27.0
